feat: Confirm before exiting application
This commit refactors the StudioWindow class in the studio.py file. A new method called confirmExit has been added to handle the confirmation dialog when the user tries to close the application. This feature enhances the functionality of the PPStudio application by providing a way for users to confirm their intention to quit the application. KeyboardInterrupts are also handled.