feat: Add mousePressed signal to CameraView
The code changes in this commit add a new signal, `mousePressed`, to the `CameraView` class. This signal is emitted when the user presses the mouse within the camera view area. The x and y coordinates of the mouse click are passed as arguments to the signal. This addition enhances the functionality of the `CameraView` class by allowing users to interact with the camera view and perform actions based on mouse clicks.